requirements.txt
matplotlib
numpy==1.24.3
pandas==2.0.3
Pillow
plotly==5.9.0
pyparsing
Requests==2.31.0
seaborn==0.13.2
streamlit==1.31.0
streamlit_extras==0.4.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.2
tensorflow
tensorflow-intel
scikit-learn
requests
tensorrt==8.5.1.7
